# Conflicts:
#	cmake/test-pkgconfig.sh
#	examples/Makefile
#	src/core.c

+++ examples/core/core_3d_camera_first_person.c
@@ -31,14 +31,14 @@ int main(void)
     camera.type = CAMERA_PERSPECTIVE;
 
     // Generates some random columns
-    float heights[MAX_COLUMNS] = { 0.0f };
+    float heights[MAX_COLUMNS] = { 0 };
     Vector3 positions[MAX_COLUMNS] = { 0 };
     Color colors[MAX_COLUMNS] = { 0 };
 
     for (int i = 0; i < MAX_COLUMNS; i++)
     {
         heights[i] = (float)GetRandomValue(1, 12);
-        positions[i] = (Vector3){ GetRandomValue(-15, 15), heights[i]/2, GetRandomValue(-15, 15) };
+        positions[i] = (Vector3){ (float)GetRandomValue(-15, 15), heights[i]/2.0f, (float)GetRandomValue(-15, 15) };
         colors[i] = (Color){ GetRandomValue(20, 255), GetRandomValue(10, 55), 30, 255 };
     }
